Move logic into hook for main screen
